move internal type to separate file
